+++ utility.h
@@ -33,4 +33,4 @@ void SaveMap(Map &dMap);
 void LoadMap(Map &dMap);
 void SaveSettings(short &fontSize, short &mapVisY);
 void LoadSettings(short &fontSize, short &mapVisY);
-double Distance(int x1, int y1, int x2, int y2);
+float Distance(int x1, int y1, int x2, int y2);
